update smoke test
