--- test/SmokeTests/OOProcSmokeTests/durableJS/DurableFunctionsHttpStart/function.json
@@ -1,24 +1,0 @@
-{
-  "bindings": [
-    {
-      "authLevel": "anonymous",
-      "name": "req",
-      "type": "httpTrigger",
-      "direction": "in",
-      "methods": [
-        "post",
-        "get"
-      ]
-    },
-    {
-      "name": "$return",
-      "type": "http",
-      "direction": "out"
-    },
-    {
-      "name": "starter",
-      "type": "orchestrationClient",
-      "direction": "in"
-    }
-  ]
-}
