--- test/SmokeTests/OOProcSmokeTests/durableJS/Hello/function.json
@@ -1,9 +1,0 @@
-{
-  "bindings": [
-    {
-      "name": "name",
-      "type": "activityTrigger",
-      "direction": "in"
-    }
-  ]
-}
