+++ test/SmokeTests/OOProcSmokeTests/durableJS/Nuget.config
@@ -4,6 +4,5 @@
         <clear/>
         <add key="nuget.org" value="https://api.nuget.org/v3/index.json"/>
         <add key="local" value="." />
-        <add key="durabletask" value="https://pkgs.dev.azure.com/durabletaskframework/734e7913-2fab-4624-a174-bc57fe96f95d/_packaging/durabletask/nuget/v3/index.json" />
     </packageSources>
 </configuration>
