--- test/SmokeTests/OOProcSmokeTests/durableJS/PrintArray/function.json
@@ -1,9 +1,0 @@
-{
-  "bindings": [
-    {
-      "name": "array",
-      "type": "activityTrigger",
-      "direction": "in"
-    }
-  ]
-}
