--- test/SmokeTests/OOProcSmokeTests/durableJS/PrintObject/function.json
@@ -1,9 +1,0 @@
-{
-  "bindings": [
-    {
-      "name": "cities",
-      "type": "activityTrigger",
-      "direction": "in"
-    }
-  ]
-}
